added a container file to FootballPitch component
